// frontend/src/api/models/syncBillRequestDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */
import type { SyncMemberCollectionDto } from "./syncMemberCollectionDto";
import type { SyncExpenseCollectionDto } from "./syncExpenseCollectionDto";

export interface SyncBillRequestDto {
  /** @nullable */
  remoteId?: string | null;
  /** @nullable */
  localId?: string | null;
  baseVersion?: number;
  /** @nullable */
  name?: string | null;
  members?: SyncMemberCollectionDto;
  expenses?: SyncExpenseCollectionDto;
  /** @nullable */
  settledTransfers?: string[] | null;
  localUpdatedAt?: string;
}