// frontend/src/api/models/syncMemberCollectionDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */
import type { SyncMemberDto } from "./syncMemberDto";

export interface SyncMemberCollectionDto {
  /** @nullable */
  upsert?: SyncMemberDto[] | null;
  /** @nullable */
  deletedIds?: string[] | null;
}