// frontend/src/api/models/syncExpenseCollectionDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */
import type { SyncExpenseDto } from "./syncExpenseDto";

export interface SyncExpenseCollectionDto {
  /** @nullable */
  upsert?: SyncExpenseDto[] | null;
  /** @nullable */
  deletedIds?: string[] | null;
}