// frontend/src/api/models/syncMemberDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */

export interface SyncMemberDto {
  /** @nullable */
  localId?: string | null;
  /** @nullable */
  remoteId?: string | null;
  /** @nullable */
  name?: string | null;
  displayOrder?: number;
  /** @nullable */
  linkedUserId?: string | null;
  /** @nullable */
  claimedAt?: string | null;
}