// frontend/src/api/models/syncExpenseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * SnapSplit API
 * 分帳計算與即時同步 API
 * OpenAPI spec version: v1
 */
import type { SyncExpenseItemCollectionDto } from "./syncExpenseItemCollectionDto";

export interface SyncExpenseDto {
  /** @nullable */
  localId?: string | null;
  /** @nullable */
  remoteId?: string | null;
  /** @nullable */
  name?: string | null;
  amount?: number;
  serviceFeePercent?: number;
  isItemized?: boolean;
  /** @nullable */
  paidByLocalId?: string | null;
  /** @nullable */
  participantLocalIds?: string[] | null;
  items?: SyncExpenseItemCollectionDto;
}